Refactor: 먹지도 조회 (#78)
- 먹지도 조회 API 선언 클래스 변경
