+++ src/test/java/dev/yogizogi/domain/meokmap/api/MeokMapApiControllerTest.java
@@ -4,7 +4,6 @@ import static dev.yogizogi.global.common.model.constant.Format.DONE;
 import static org.mockito.ArgumentMatchers.eq;
 import static org.mockito.BDDMockito.given;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.delete;
-import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.post;
 import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.print;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.content;
@@ -13,18 +12,13 @@ import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import dev.yogizogi.domain.meokmap.factory.dto.AddRestaurantOnMapFactory;
-import dev.yogizogi.domain.meokmap.factory.dto.RetrieveMeokMapFactory;
 import dev.yogizogi.domain.meokmap.factory.dto.RemoveRestaurantOnMapFactory;
-import dev.yogizogi.domain.meokmap.factory.fixtures.MeokMapFixtures;
 import dev.yogizogi.domain.meokmap.factory.fixtures.MeokMapRestaurantFixtures;
 import dev.yogizogi.domain.meokmap.model.dto.request.AddRestaurantOnMeokMapInDto;
 import dev.yogizogi.domain.meokmap.model.dto.request.RemoveRestaurantOnMeokMapInDto;
-import dev.yogizogi.domain.meokmap.model.dto.response.RetrieveMeokMapOutDto;
 import dev.yogizogi.domain.meokmap.service.MeokMapService;
 import dev.yogizogi.domain.security.service.JwtService;
 import java.nio.charset.StandardCharsets;
-import java.util.List;
-import java.util.UUID;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -40,9 +34,9 @@ import org.springframework.test.web.servlet.MockMvc;
 
 @ActiveProfiles("test")
 @WebMvcTest(value = {MeokMapApiController.class},
-    excludeFilters =  {
-            @ComponentScan.Filter(type = FilterType.ASSIGNABLE_TYPE, classes = SecurityConfig.class)
-    }
+        excludeFilters =  {
+                @ComponentScan.Filter(type = FilterType.ASSIGNABLE_TYPE, classes = SecurityConfig.class)
+        }
 )
 @AutoConfigureMockMvc(addFilters = false)
 @DisplayName("MeokMapApiController 동작 테스트")
@@ -118,41 +112,5 @@ class MeokMapApiControllerTest {
                 );
 
     }
-
-    @Test
-    void 먹지도_조회() throws Exception {
-
-        // given
-        UUID 사용자 = MeokMapFixtures.사용자.getId();
-        List<RetrieveMeokMapOutDto> 응답 = RetrieveMeokMapFactory.retrieveMeokMapOutDtos();
-
-        // mocking
-        given(jwtService.getUserId())
-                .willReturn(사용자);
-
-        given(meokMapService.retrieveMeokMap(eq(사용자)))
-                .willReturn(응답);
-
-        // when
-        // then
-        mockMvc.perform(
-                        get("/api/meok-maps")
-                                .contentType(MediaType.APPLICATION_JSON)
-                                .characterEncoding(StandardCharsets.UTF_8)
-                )
-                .andDo(print())
-                .andExpect(status().isOk())
-                .andExpect(
-                        content().contentTypeCompatibleWith(MediaType.APPLICATION_JSON)
-                )
-                .andExpect(
-                        jsonPath("$.data[0].restaurantId").value(응답.get(0).getRestaurantId().toString())
-                );
-
-    }
-
-
-
-
 
 }
